admin/app/requirements.txt
django-constance[redis]
django-environ
django-extensions
Django>=3.0,<4.0
gunicorn
psycopg2-binary>=2.8
python-dotenv
supervisor
